requirements.txt
# 爬虫与反爬
selenium>=4.10.0
undetected-chromedriver>=3.5.0
selenium-stealth>=1.0.6
webdriver-manager>=4.0.0
fake-useragent>=1.1.1
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pyautogui>=0.9.54
Pillow>=10.0.0

# 数据处理与存储
pandas>=2.0.3
numpy>=1.24.3
openpyxl>=3.1.2
xlsxwriter>=3.1.0

# 文本分析
scikit-learn>=1.3.0
nltk>=3.8.1
spacy>=3.6.1

# LLM 抽取
google-generativeai>=0.3.0
tenacity>=8.2.3
tiktoken>=0.5.1
pydantic>=2.4.2

# 可视化
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0
wordcloud>=1.9.2

# 交互界面
gradio>=4.0.0

# 辅助工具
tqdm>=4.66.1
logging>=0.5.1.2
retrying>=1.3.4
schedule>=1.2.0
apscheduler>=3.10.1